requirements.txt
pandas
openpyxl
sentence-transformers
faiss-cpu
streamlit
python-dotenv